Fix repeated generators bug.
Our Schreier-Sims works well enough to make all of the previous tests
pass (albeit somewhat more slowly). The problem was in the
initialisation of the base at the start of the procedure (the function
ExtendBaseIfStabilized). Previously this took the case when it was
called with an empty basis as a special one, and added only one element.
I'm not sure why, but I think this must have been an attempt to work
around a bug elsewhere and was not removed.

Even though Strip does not require a full BSGS in our case, it does
require enough base elements for the stabiliser we are computing. It
never got a chance to try all the generators, and so incorrectly
decided that the generators were not in the group when passed them.