Fix a bug related to basic orbit computation
Basically if the largest point a stabiliser group acted on was less than the largest point acted on by the whole group (I'm surprised this didn't come up earlier...), then the Schreier vector for the corresponding orbit is made a shorter list. But then occasionally we query it for those larger points --- we know they're not in the orbit, but asking the question leads to an out of bounds access. So for now, pad the Schreier vector in this situation.

The random group which acted as a test case for this was TransitiveGroup(10, 37) = [2^4]S(5) in GAP's notation.