Remove sanity check from hot function
We previously checked to make sure we didn't get into an infinite loop in SchreierVectorPermFromBasePoint, except this function is called *a lot*, and now I am a little more confident that the Schreier vector computations are correct. Anyway, an infinite loop running the tests will give more of an indication of the probem.