Fix group order check in RandomSchreierSims (& small optimisations)
A silly typo in RandomSchreierSims meant that we never got to bail out early in the simple case where we knew the order of a group a priori. This gives a big speedup in this case (which happens to be most of our tests because of how we pick the groups).

Other small optimisations include moving some checks and removing some that are no longer necessary.