Finally fix point-swap change of base
The issue was not the point swapping but mostly a rather overzealous
cleaning up of the 'tail' of potentially old base points left over at
the end. Now if we are given a base with repeated entries or which
finishes with redundant points, we still get a valid chain out with the
specified base.