Fix bug in GAPStabChain code with trivial chain entries
If we have (as a result of, say, a change of base) trivial chain entries — or in fact any entry where () is in stabgens — then we (correctly) can't find them in the SGS, since we don't put () there usually. This isn't a problem, since GAP always keeps () in its labels list, but we need to special-case the index assignment.