Cache stabilizer chains as they are computed
When a change of base is requested, we consult a (currently unbounded)
cache; this takes the form of a prefix tree with a configurable depth
limit. As part of this refactoring, the base change API has been
extended; there are now three calls:
  BSGSWithBase(bsgs, base) - the primary interface
  BSGSWithBasePrefix(bsgs, prefix) - when only initial segments are
    important.
  ChangeBaseOfBSGS(bsgs, base) - an in-place mutating change of base.
To take advantage of the new surface, and to test it, a method has been
added to compute pointwise stabilisers.