Work around some odd issues with code coverage reporting
There are lines which definitely should be skipped which aren't in the
code coverage report (they are in some files, but not in others). Whose
responsibility this is, I'm not sure, but the headline figure is rather
low (even in comparison to GAP), so I've bodged a script to fix it up
manually.

(We also hopefully improve coverage a little by making sure we hit the
removal code in the LRU chain cache.)