Test and fix a broken base change case
Previously, we did not test the case where a base point to be introduced
was already stabilized by an existing nontrivial stabilizer group --- in
which case it is inserted earlier on in the base to reduce the number of
swaps required. Code is added to create new bases like this when
testing. Exercising this code path also revealed a related off-by-one
error, which is fixed.